Logistic Regression
Logistic Regression
This project involves creating a model that will predict whether or not a user will click on an ad-based off the features of that user.

The data set being used contains the following features:

'Daily Time Spent on Site': consumer time on-site in minutes
'Age': customer age in years
'Area Income': Avg. The income of the geographical area of consumer
'Daily Internet Usage': Avg. minutes a day consumer is on the internet
'Ad Topic Line': Headline of the advertisement
'City': City of consumer
'Male': Whether or not consumer was male
'Country': Country of consumer
'Timestamp': Time at which consumer clicked on Ad or closed window
'Clicked on Ad': 0 or 1 indicated clicking on Ad